+++ functions/src/middlewares/csrf_middleware.ts
@@ -5,6 +5,7 @@ import * as functions from "firebase-functions";
 const csrfExemptRoutes = [
   "/auth/login",
   "/auth/register",
+  "/auth/session-login",
   // "/auth/reset-password",
 ]
 
@@ -23,7 +24,7 @@ export const csrfProtection = (req: Request, res: Response, next: NextFunction) 
 
   if (!csrfCookie || !csrfHeader || csrfCookie !== csrfHeader) {
     functions.logger.log("CSRF validation rejected as cookie and header does not match.")
-    return res.status(403).json({ error: "CSRF token validation failed" });
+    return res.status(403).json({ status_code: 403, error: "CSRF token validation failed" });
   }
 
   return next();
